requirements.txt
tweepy>=4.16.0
python-dotenv>=1.1.1
pandas>=2.3.2
tqdm>=4.67.1
